requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pymongo
pandas
faiss-cpu
tqdm
typer
python-dotenv
